9.7.20
Changed the way in which Cysts are deleted to try and get rid of Exploiting. It will definitely effect/affect (whatever the freakin word is) ALL maps. Previously the Cyst OnInitialized function would check if it's by a FrontDoor within 13 radius. Now the FrontDoor checks if not opened every 1 second for a Cyst within 17 radius.
BigMac OnOrderCompleted - added if checks if self.autoReturning or currentOrder:GetType() == kTechId.Move then .... if currentOrder:GetType() == kTechId.Move then ....